# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: CDX Loap Huong Notification Service API
  description: |
    Comprehensive notification service API for CDX Loap Huong platform.

    ## Features
    - Device token management for push notifications
    - Notification history and management
    - User notification preferences
    - Admin broadcast capabilities
    - Retry mechanism for failed notifications
    - Real-time delivery tracking

    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.

    ## Rate Limiting
    - User endpoints: 1000 requests per minute
    - Admin endpoints: 100 requests per minute

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: CDX Development Team
    email: dev@cdx.com
  license:
    name: Proprietary
    url: https://cdx.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.cdx.com
    description: Production server
  - url: https://staging-api.cdx.com
    description: Staging server
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notification history
      description: Retrieve notification history with pagination and filtering
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: type
          in: query
          description: Filter by notification type
          schema:
            $ref: '#/components/schemas/NotificationType'
        - name: channel
          in: query
          description: Filter by notification channel
          schema:
            $ref: '#/components/schemas/NotificationChannel'
        - name: status
          in: query
          description: Filter by notification status
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: startDate
          in: query
          description: Filter notifications from this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter notifications until this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, sentAt, readAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Notification history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /notifications/{id}:
    get:
      tags:
        - Notifications
      summary: Get notification detail
      description: Retrieve detailed information about a specific notification
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
            example: 'notif_123456789'
      responses:
        '200':
          description: Notification detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
            example: 'notif_123456789'
      responses:
        '200':
          description: Notification marked as read successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAsReadResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Mark all notifications as read
      description: Mark all unread notifications for the user as read
      responses:
        '200':
          description: All notifications marked as read successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllAsReadResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /notifications/unread-count:
    get:
      tags:
        - Notifications
      summary: Get unread count
      description: Get the count of unread notifications for the user
      responses:
        '200':
          description: Unread count retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # User Preferences
  /preferences:
    get:
      tags:
        - Preferences
      summary: Get user notification preferences
      description: Retrieve notification preferences for the authenticated user
      responses:
        '200':
          description: Preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Preferences
      summary: Update user notification preferences
      description: Update notification preferences for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
            examples:
              full_preferences:
                summary: Complete Preferences
                value:
                  push:
                    enabled: true
                    types: ['payment', 'booking', 'announcement']
                    quietHours:
                      enabled: true
                      start: '22:00'
                      end: '08:00'
                      timezone: 'Asia/Ho_Chi_Minh'
                  email:
                    enabled: true
                    types: ['announcement', 'system']
                    frequency: 'daily'
                  sms:
                    enabled: false
                    types: []
              minimal_preferences:
                summary: Minimal Preferences
                value:
                  push:
                    enabled: true
                    types: ['payment']
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Admin Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  description: Error code
                details:
                  type: string
                  description: Detailed error message
                field:
                  type: string
                  description: Field that caused the error (for validation errors)
...
    Notification:
      type: object
      required:
        - id
        - userId
        - title
        - body
        - type
        - channel
        - status
        - createdAt
      properties:
        id:
          type: string
          description: Unique notification ID
        userId:
          type: string
          description: User ID
        notificationId:
          type: string
          description: Original notification ID
        title:
          type: string
          description: Notification title
        body:
          type: string
          description: Notification body
        type:
          $ref: '#/components/schemas/NotificationType'
        channel:
          $ref: '#/components/schemas/NotificationChannel'
        priority:
          $ref: '#/components/schemas/NotificationPriority'
        status:
          $ref: '#/components/schemas/NotificationStatus'
        data:
          type: object
          description: Additional notification data
        sentAt:
          type: string
          format: date-time
          description: When the notification was sent
        deliveredAt:
          type: string
          format: date-time
          description: When the notification was delivered
        readAt:
          type: string
          format: date-time
          description: When the notification was read
        errorMessage:
          type: string
          description: Error message if notification failed
        errorCode:
          type: string
          description: Error code if notification failed
        retryCount:
          type: integer
          description: Number of retry attempts
        deliveryId:
          type: string
          description: External delivery service ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NotificationHistoryResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - notifications
                - pagination
              properties:
                notifications:
                  type: array
                  items:
                    $ref: '#/components/schemas/Notification'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    NotificationDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Notification'
    MarkAsReadResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - success
                - notificationId
                - readAt
              properties:
                success:
                  type: boolean
                notificationId:
                  type: string
                readAt:
                  type: string
                  format: date-time
    MarkAllAsReadResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - success
                - updatedCount
                - readAt
              properties:
                success:
                  type: boolean
                updatedCount:
                  type: integer
                readAt:
                  type: string
                  format: date-time
    UnreadCountResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - count
                - lastUpdated
              properties:
                count:
                  type: integer
                  description: Number of unread notifications
                lastUpdated:
                  type: string
                  format: date-time
                  description: When the count was last updated

    # Preferences Schemas
    NotificationPreferences:
      type: object
      properties:
        push:
          $ref: '#/components/schemas/ChannelPreferences'
        email:
          $ref: '#/components/schemas/ChannelPreferences'
        sms:
          $ref: '#/components/schemas/ChannelPreferences'
    ChannelPreferences:
      type: object
      required:
        - enabled
        - types
      properties:
        enabled:
          type: boolean
          description: Whether this channel is enabled
        types:
          type: array
          items:
            $ref: '#/components/schemas/NotificationType'
          description: Allowed notification types for this channel
        quietHours:
          type: object
          properties:
            enabled:
              type: boolean
            start:
              type: string
              pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
              example: '22:00'
            end:
              type: string
              pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
              example: '08:00'
            timezone:
              type: string
              example: 'Asia/Ho_Chi_Minh'
        frequency:
          type: string
          enum: [immediate, hourly, daily, weekly]
          description: Email frequency preference
    UpdatePreferencesRequest:
      type: object
      properties:
        push:
          $ref: '#/components/schemas/ChannelPreferences'
        email:
          $ref: '#/components/schemas/ChannelPreferences'
        sms:
          $ref: '#/components/schemas/ChannelPreferences'
    PreferencesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - userId
                - preferences
                - createdAt
                - updatedAt
              properties:
                userId:
                  type: string
                preferences:
                  $ref: '#/components/schemas/NotificationPreferences'
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time

    # Admin Schemas
...
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
      properties:
        page:
          type: integer
          description: Current page number (1-based)
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page

    # Enum Schemas
...
    NotificationType:
      type: string
      enum: [payment, booking, announcement, system, marketing, security]
      description: Type of notification
    NotificationChannel:
      type: string
      enum: [push, email, sms, inApp]
      description: Notification delivery channel
    NotificationPriority:
      type: string
      enum: [low, normal, high, urgent]
      description: Notification priority level
    NotificationStatus:
      type: string
      enum: [pending, sent, delivered, read, failed, cancelled]
      description: Notification status
  responses:
    BadRequestError:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Validation failed'
            error:
              code: 'VALIDATION_ERROR'
              details: 'token is required'
              field: 'token'
            timestamp: '2024-01-15T10:30:00Z'
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Unauthorized'
            error:
              code: 'UNAUTHORIZED'
              details: 'Invalid or missing JWT token'
            timestamp: '2024-01-15T10:30:00Z'
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Forbidden'
            error:
              code: 'FORBIDDEN'
              details: 'Admin access required'
            timestamp: '2024-01-15T10:30:00Z'
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Not found'
            error:
              code: 'NOT_FOUND'
              details: 'Notification not found'
            timestamp: '2024-01-15T10:30:00Z'
...